androidmedia: Fix stagnation when obtaining thumb
vs-thumb was never finishing when trying to get a thumbnail for very
short videos. This happened because EOS was received but the decoding
thread did not signal the 'drained' condition, which is fixed by this
commit. Note that in some cases you might not get the thumbnail anyway,
but we make sure that vs-thumb stops.