androidmedia: Avoid deadlock when seeking
In some cases a deadlock was arising when a seek was being executed.
This happened because an infinite loop in gst_amc_video_dec_handle_frame
was preventing the pad stream lock from being freed. This is now avoided
so the streaming thread does not get stuck in this function.